refactor(admin): split dashboard route modules
Why:
- keep the Quantercise dashboard route as a shell instead of mixing fetch, fallback, and panel rendering
- split dashboard metrics and activity panels behind the existing dashboard-panels export surface

Verify:
- pnpm exec prettier --check apps/admin/src/app/quantercise/page.tsx apps/admin/src/app/quantercise/dashboard-panels.tsx apps/admin/src/app/quantercise/dashboard-content.tsx apps/admin/src/app/quantercise/dashboard-skeleton.tsx apps/admin/src/app/quantercise/dashboard-metric-panels.tsx apps/admin/src/app/quantercise/dashboard-activity-panels.tsx
- git diff --check -- apps/admin/src/app/quantercise/page.tsx apps/admin/src/app/quantercise/dashboard-panels.tsx apps/admin/src/app/quantercise/dashboard-content.tsx apps/admin/src/app/quantercise/dashboard-skeleton.tsx apps/admin/src/app/quantercise/dashboard-metric-panels.tsx apps/admin/src/app/quantercise/dashboard-activity-panels.tsx
- pnpm --filter @anipotts/admin typecheck
- pnpm --filter @anipotts/admin lint
- pnpm --filter @anipotts/admin build
- pnpm validate
