+++ apps/admin/src/app/quantercise/dashboard-skeleton.tsx
@@ -1,0 +1,14 @@
+import { PanelSkeleton } from "./components";
+
+export function DashboardSkeleton() {
+  return (
+    <>
+      <PanelSkeleton title="Key Metrics" />
+      <PanelSkeleton title="Alerts" />
+      <PanelSkeleton title="Quick Stats" />
+      <PanelSkeleton title="Recent Activity" />
+      <PanelSkeleton title="Content Pipeline" />
+      <PanelSkeleton title="Last Updated" />
+    </>
+  );
+}
